Refactored and commented card holders
