+++ Assets/Scripts/CardHolders/CardArea.cs
@@ -17,27 +17,9 @@ public abstract class CardArea : CardHolder
         heldCards = new CardBase[6];
     }
 
-    public virtual void AddCard(CardBase card)
+    public override void RemoveCard(CardBase card)
     {
-        if (!card) { return; }
-        if (currentHeldCount >= heldCards.Length) { return; }
-        cardXStartPos = cardXPosStarts[currentHeldCount];
-        heldCards[Array.IndexOf(heldCards, null)] = card;
-        currentHeldCount++;
-        ReorganizeHeldCardPositions();
-    }
-
-    public virtual void RemoveCard(CardBase card)
-    {
-        if (!card) { return; }
-        if (!heldCards.Contains(card)) { return; }
-
-        var index = Array.IndexOf(heldCards, card);
-        heldCards[index] = null;
-        Destroy(card.gameObject);
-        currentHeldCount--;
-        cardXStartPos = cardXPosStarts[Mathf.Clamp(currentHeldCount - 1, 0, cardXPosStarts.Length - 1)];
-        ReorganizeHeldCardPositions();
+        base.RemoveCard(card);
 
         if (currentHeldCount <= 0)
         {
